# specs/009-post-creation/contracts/api.yaml
openapi: 3.0.0
info:
  title: Post Creation API
  version: 1.0.0
  description: API for post creation and interactions

paths:
  /api/posts:
    get:
      summary: Get user feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 500
                media:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [image, video, link]
                      file:
                        type: string
                      url:
                        type: string
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: integer

  /api/posts/{id}/like:
    post:
      summary: Toggle like on post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  liked:
                    type: boolean
                  count:
                    type: integer

  /api/posts/{id}/comment:
    post:
      summary: Add comment to post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 200
                parentCommentId:
                  type: integer
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                type: object
                properties:
                  commentId:
                    type: integer

  /api/posts/{id}/share:
    post:
      summary: Share post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post shared
          content:
            application/json:
              schema:
                type: object
                properties:
                  shareUrl:
                    type: string

  /api/posts/{id}/save:
    post:
      summary: Toggle save on post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved:
                    type: boolean

components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
        likes:
          type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"

    Comment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"